feat(apps): add accounting app and move models
This commit introduces a new `accounting` app and moves the
following models from the `workflow` app to the new app:

- `Bill`
- `BillLineItem`
- `CreditNote`
- `CreditNoteLineItem`
- `Invoice`
- `InvoiceLineItem`
- `Quote`

This improves the organization of the codebase by grouping
related models together.  The `XeroJournal` and
`XeroJournalLineItem` models remain in the `workflow` app
due to their close relationship with Xero integration
logic.
